+++ admin/src/editor/conversion.tsx
@@ -1,11 +1,23 @@
 /* eslint-disable react/display-name */
 
-import React, { FC, ReactNode } from "react";
-import { Text, Element as SlateElement } from "slate";
+import React, { FC, ReactNode, ReactChild } from "react";
+import { renderToStaticMarkup } from "react-dom/server";
+import {
+  Text as SlateText,
+  Element as SlateElement,
+  Node as SlateNode,
+} from "slate";
 import { jsx } from "slate-hyperscript";
-import { Mark, Block, BlockFormat, BlockTagType } from "./formats";
+import {
+  Mark,
+  Block,
+  BlockFormat,
+  BlockTagType,
+  isLinkElement,
+} from "./formats";
+import { flatMap } from "../flat-map";
 
-export const SerializedMark: FC<{ readonly leaf: Text }> = ({
+export const SerializedMark: FC<{ readonly leaf: SlateText }> = ({
   children,
   leaf,
 }) => {
@@ -25,8 +37,8 @@ export const SerializedMark: FC<{ readonly leaf: Text }> = ({
   return <>{ret}</>;
 };
 
-export const deserializeMark = (element: Element): Text => {
-  const slateNode: Text = { text: element.textContent ?? "" };
+const deserializeMark = (element: Element): SlateText => {
+  const slateNode: SlateText = { text: element.textContent ?? "" };
   switch (element.tagName) {
     case "CODE":
       slateNode[Mark.Code] = true;
@@ -82,7 +94,7 @@ interface DeserializerProps {
 
 type Deserialzier = (props: DeserializerProps) => ReturnType<typeof jsx>;
 
-export const blockDeserializers: Record<BlockTagType, Deserialzier> = {
+const blockDeserializers: Record<BlockTagType, Deserialzier> = {
   A: ({ element, children }) =>
     jsx(
       "element",
@@ -108,5 +120,66 @@ export const blockDeserializers: Record<BlockTagType, Deserialzier> = {
     ),
 };
 
-export const isDeserializable = (x: string): x is BlockTagType =>
+const isDeserializable = (x: string): x is BlockTagType =>
   x in blockDeserializers;
+
+const serializeAsHTML = (node: SlateNode): ReactChild | null => {
+  if (SlateText.isText(node)) {
+    return <SerializedMark leaf={node}>{node.text}</SerializedMark>;
+  }
+
+  const children = node.children.map(n => serializeAsHTML(n));
+
+  if (isLinkElement(node)) {
+    return <a href={node.url}>{children}</a>;
+  }
+
+  const blockSerializer = blockSerializers[node["type"] as BlockFormat];
+  if (blockSerializer) {
+    return blockSerializer({ children, element: node });
+  }
+
+  return <>{children}</>;
+};
+
+export const serialize = (nodes: SlateNode[]): string =>
+  renderToStaticMarkup(<>{nodes.map(n => serializeAsHTML(n))}</>);
+
+const isTextNode = (node: Node): node is Text =>
+  node.nodeType === Node.TEXT_NODE;
+
+const isElementNode = (node: Node): node is Element =>
+  node.nodeType === Node.ELEMENT_NODE;
+
+const textElement = (node: Node): SlateText => ({
+  text: node.textContent ?? "",
+});
+
+const deserialize = (el: Node): Array<ReturnType<typeof jsx>> => {
+  if (isTextNode(el)) {
+    return [textElement(el)];
+  }
+
+  if (isElementNode(el)) {
+    const children = flatMap(Array.from(el.childNodes), deserialize);
+
+    const { nodeName } = el;
+    if (isDeserializable(nodeName)) {
+      return [blockDeserializers[nodeName]({ children, element: el })];
+    }
+
+    switch (el.nodeName) {
+      case "BODY":
+        return jsx("fragment", {}, children);
+      case "BR":
+        return [{ text: "\n" }];
+      default:
+        return [deserializeMark(el)];
+    }
+  }
+
+  return [];
+};
+
+export const deserializeHTML = (input: string): any =>
+  deserialize(new DOMParser().parseFromString(input, "text/html").body);
